Rename "warc" app to "crawler"
For clarity, this renames the existing Django "warc" app to "crawler".
This means this project now has two Django apps: "crawler" and
"viewer". This should clarify the function of each app.

This change impacts the database schema so, after merging, the
database will need to be regenerated.
