+++ settings.py
@@ -37,8 +37,8 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     "django_filters",
     "rest_framework",
+    "crawler",
     "viewer",
-    "warc",
 ]
 
 MIDDLEWARE = [
